+++ packages/blade/src/components/Box/__tests__/Box.web.test.tsx
@@ -17,6 +17,11 @@ describe('<Box />', () => {
         maxHeight={56}
         minWidth={48}
         maxWidth={56}
+        position="absolute"
+        top={0}
+        right={0}
+        bottom={0}
+        left={0}
       />,
     );
     expect(container).toMatchSnapshot();