+++ packages/blade/src/storybook-recipes/AccessibilityInterop/AccessibilityInteropDemo.web.tsx
@@ -2,12 +2,15 @@ import React from 'react';
 import { getCheckboxAccessibilityProps } from './getCheckboxA11yProps';
 import { makeAccessible } from '~utils';
 import { CloseIcon, EyeIcon } from '~components/Icons';
+import Box from '~components/Box';
+import { useTheme } from '~components/BladeProvider';
 
 const Checkbox: React.FC<{ label: string; checked?: boolean; disabled?: boolean }> = ({
   label,
   checked,
   disabled,
 }) => {
+  const { theme } = useTheme();
   const [isChecked, setChecked] = React.useState(checked);
   const checkboxAccessibilityProps = getCheckboxAccessibilityProps({
     isDisabled: disabled,
@@ -24,23 +27,26 @@ const Checkbox: React.FC<{ label: string; checked?: boolean; disabled?: boolean 
   };
 
   return (
-    <>
-      <span>
-        {isChecked ? (
-          <EyeIcon color="action.icon.link.active" size="small" />
-        ) : (
-          <CloseIcon color="action.icon.link.active" size="small" />
-        )}
-      </span>
-      <span
+    <Box display="flex" alignItems="center" gap="spacing.2">
+      {isChecked ? (
+        <EyeIcon color="action.icon.link.active" size="medium" />
+      ) : (
+        <CloseIcon color="action.icon.link.active" size="medium" />
+      )}
+      <Box
         onKeyDown={handleOnKeyDown}
         onClick={toggleChecked}
         {...checkboxAccessibilityProps}
         tabIndex={0}
+        style={{
+          color: disabled
+            ? theme.colors.action.text.primary.disabled
+            : theme.colors.surface.text.normal.lowContrast,
+        }}
       >
         {label}
-      </span>
-    </>
+      </Box>
+    </Box>
   );
 };
 
@@ -53,8 +59,9 @@ const AccessibilityInteropDemo = (): React.ReactElement => {
   return (
     <>
       <h3 id="id-group-label">Fruits</h3>
+      <Box marginBottom="spacing.1" />
       <div {...checkboxGroupA11y}>
-        <ul>
+        <ul style={{ listStyle: 'none', margin: 0, padding: 0 }}>
           <li>
             <Checkbox checked label="Mango" />
           </li>