+++ packages/blade/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @razorpay/blade
 
+## 0.13.6
+
+### Patch Changes
+
+- b365464: fix: button spinner layout
+- f3abfbe: feat(Icons): add new icons
+
+## 0.13.5
+
+### Patch Changes
+
+- 7909d7c: fix(blade): Checkbox design changes
+
 ## 0.13.4
 
 ### Patch Changes